Note: Uninstall app to refresh DB! -Initialized the "assigned catererstaff" db and created a dummie 1st entry

+++ app/src/main/java/comcse5324projutacatering/httpsgithub/utacatering/ViewEditProfile.java
@@ -70,7 +70,7 @@ public class ViewEditProfile extends Activity {
 
                 String username = editUsername.getText().toString();
                 String password = editPassword.getText().toString();
-                int    stu_id   = Integer.parseInt(editStudentID.getText().toString());
+                long    stu_id   = Long.parseLong(editStudentID.getText().toString());
                 String contact  = editContactDetails.getText().toString();
                 String personal = editPersonalDetails.getText().toString();
